+++ Services/IQuestService.cs
@@ -1,0 +1,7 @@
+﻿namespace SimpleBackendGame.Services
+{
+    public interface IQuestService
+    {
+        bool GoQuest(int questId);
+    }
+}
